macprefs/command: simplify generation of domains

+++ macprefs/command.py
@@ -11,75 +11,42 @@ import asyncio.subprocess as sp
 import plistlib
 import sys
 
-from .constants import (DOMAIN_SPLIT_DELIMITER, GLOBAL_DOMAIN_ARG,
-                        MAX_CONCURRENT_EXPORT_TASKS)
+from .constants import GLOBAL_DOMAIN_ARG, MAX_CONCURRENT_EXPORT_TASKS
 from .filters import BAD_DOMAINS
 from .mp_typing import PlistRoot
 from .plist2defaults import plist_to_defaults_commands
 from .processing import remove_data_fields
-from .shell import delete_old_plists, git
+from .shell import git
 from .utils import setup_logging_stderr
 
 __all__ = ('main', )
 
 
-async def _generate_domains(repo_prefs_dir: Path,
-                            out_dir: str) -> AsyncIterator[str]:
-    log = setup_logging_stderr(verbose=True)
-    prefs_dir = Path.home().joinpath('Library/Preferences')
-    can_skip: List[int] = []
-    p = await asyncio.create_subprocess_shell('defaults domains',
-                                              stdout=sp.PIPE)
-    deletions = []
-    assert p.stdout is not None
-    for i, domain in enumerate((await p.stdout.read()).decode('utf-8').replace(
-            ',', '').split(DOMAIN_SPLIT_DELIMITER)):
-        if (len(domain.strip()) == 0
-                or domain == '$(PRODUCT_BUNDLE_IDENTIFIER)'):
-            deletions.append(domain)
-            continue
-        if not prefs_dir.joinpath(f'{domain}.plist').exists():
-            p = await asyncio.create_subprocess_shell(
-                f'defaults read {quote(domain)}',
-                stdout=sp.PIPE,
-                stderr=sp.PIPE)
-            await p.wait()
-            if p.returncode != 0:
-                candidates = list(prefs_dir.glob(f'{domain}*'))
-                if not candidates:
-                    if i in can_skip:
-                        deletions.append(domain)
-                        continue
-                    assert p.stderr is not None
-                    if b'does not exist' in await p.stderr.read():
-                        log.info('"%s" listed but does not exist',
-                                 domain.strip())
-                        deletions.append(domain)
-                        continue
-                    raise ValueError(f'Not sure how to handle {domain}')
-                if len(candidates) != 1:
-                    candidates_joined = ', '.join(x.name for x in candidates)
-                    raise ValueError(
-                        f'Multiple candidates: {candidates_joined}')
-                domain = candidates[0].stem
-                can_skip.extend(
-                    range(i + 1,
-                          i + len(domain.split(DOMAIN_SPLIT_DELIMITER))))
-                yield domain
-        else:
-            yield domain
+async def _has_git() -> bool:
+    p = await sp.create_subprocess_shell('bash -c "command -v git"',
+                                         stdout=sp.PIPE)
+    await p.wait()
+    return p.returncode == 0
+
+
+async def _generate_domains() -> AsyncIterator[str]:
+    for plist in (x for x in (
+            Path.home().joinpath('Library/Preferences').glob('*.plist'))
+                  if x.stem and x.stem != '$(PRODUCT_BUNDLE_IDENTIFIER)'
+                  and not x.name.startswith('.')):
+        yield plist.stem
     yield GLOBAL_DOMAIN_ARG
-    await delete_old_plists(deletions, repo_prefs_dir, work_tree=out_dir)
 
 
 async def _defaults_export(domain: str,
-                           repo_prefs_dir: Path) -> Tuple[str, PlistRoot]:
+                           repo_prefs_dir: Path,
+                           debug: bool = False) -> Tuple[str, PlistRoot]:
     command = f'defaults export {quote(domain)}'
     out_domain = 'globalDomain' if domain == GLOBAL_DOMAIN_ARG else domain
     plist_out = repo_prefs_dir.joinpath(f'{out_domain}.plist')
     path_quoted = quote(str(plist_out))
     command += f' {path_quoted}'
-    log = setup_logging_stderr(verbose=True)
+    log = setup_logging_stderr(verbose=debug)
     log.debug('Running: %s', command)
     p = await asyncio.create_subprocess_shell(command,
                                               stdout=sp.PIPE,
@@ -105,19 +72,21 @@ async def _setup_out_dir(out_dir: str) -> Tuple[str, Path]:
     return out_dir, repo_prefs_dir
 
 
-async def _main(out_dir: str) -> int:
-    out_dir, repo_prefs_dir = await _setup_out_dir(out_dir)
+async def _main(out_dir: str, debug: bool = False) -> int:
+    log = setup_logging_stderr(verbose=debug)
+    has_git = await _has_git()
 
+    out_dir, repo_prefs_dir = await _setup_out_dir(out_dir)
     export_tasks = []
     all_data: List[Tuple[str, PlistRoot]] = []
-    async for domain in _generate_domains(repo_prefs_dir, out_dir):
+    async for domain in _generate_domains():
         # spell-checker: disable
         if domain in ('com.apple.Music', 'com.apple.TV',
                       'com.apple.identityservices.idstatuscache',
                       'com.apple.security.KCN'):
             # spell-checker: enable
             continue
-        export_tasks.append(_defaults_export(domain, repo_prefs_dir))
+        export_tasks.append(_defaults_export(domain, repo_prefs_dir, debug))
         if len(export_tasks) == MAX_CONCURRENT_EXPORT_TASKS:
             all_data.extend(await asyncio.gather(*export_tasks))
             export_tasks = []
@@ -133,36 +102,40 @@ async def _main(out_dir: str) -> int:
         for domain, root in sorted(all_data, key=lambda x: x[0]):
             if not root:
                 continue
-            async for line in plist_to_defaults_commands(domain, root):
+            async for line in plist_to_defaults_commands(domain, root, debug):
                 f.write(line + '\n')
             out_domain = ('globalDomain'
                           if domain == GLOBAL_DOMAIN_ARG else domain)
             known_domains.append(out_domain)
             plist_path = repo_prefs_dir.joinpath(f'{out_domain}.plist')
-            p = await asyncio.create_subprocess_shell(
-                f'plutil -convert xml1 {quote(str(plist_path))}')
+            cmd = f'plutil -convert xml1 {quote(str(plist_path))}'
+            log.debug('Executing: %s', cmd)
+            p = await asyncio.create_subprocess_shell(cmd)
             tasks.append(p.wait())
     chmod(str(exec_defaults), 0o755)
-
     await asyncio.wait(tasks)
 
-    log = setup_logging_stderr(verbose=True)
-
-    # Clean up very old plists
-    delete_with_git = (str(j[1]) for j in (
-        (file, file_) for file, file_ in ((x, repo_prefs_dir.joinpath(x))
+    if has_git:
+        # Clean up very old plists
+        delete_with_git = [
+            str(j[1])
+            for j in ((file, file_)
+                      for file, file_ in ((x, repo_prefs_dir.joinpath(x))
                                           for x in listdir(str(repo_prefs_dir))
                                           if x != '.gitignore')
-        if file[:-6] not in known_domains and file_.exists()
-        if not file_.is_dir()))
-    await git(chain(('rm', '-f', '--ignore-unmatch', '--'), delete_with_git),
-              check=True,
-              work_tree=out_dir)
-    all_files = ' '.join(map(quote, delete_with_git))
-    cmd = f'rm -f -- {all_files}'
-    log.debug('Executing: %s', cmd)
-    p = await asyncio.create_subprocess_shell(f'rm -f -- {all_files}')
-    await p.wait()
+                      if file[:-6] not in known_domains and file_.exists()
+                      if not file_.is_dir())
+        ]
+        await git(chain(('rm', '-f', '--ignore-unmatch', '--'),
+                        delete_with_git),
+                  check=True,
+                  work_tree=out_dir,
+                  debug=debug)
+        all_files = ' '.join(map(quote, delete_with_git))
+        cmd = f'rm -f -- {all_files}'
+        log.debug('Executing: %s', cmd)
+        p = await asyncio.create_subprocess_shell(f'rm -f -- {all_files}')
+        await p.wait()
 
     delete_with_git_ = []
     delete_with_rm = []
@@ -172,9 +145,12 @@ async def _main(out_dir: str) -> int:
         plist = repo_prefs_dir.joinpath(f'{x}.plist')
         delete_with_git_.append(str(plist))
         delete_with_rm.append(quote(str(plist)))
-    await git(['rm', '-f', '--ignore-unmatch', '--'] + delete_with_git_,
-              check=True,
-              work_tree=out_dir)
+    if has_git:
+        await git(chain(('rm', '-f', '--ignore-unmatch', '--'),
+                        delete_with_git_),
+                  check=True,
+                  debug=debug,
+                  work_tree=out_dir)
     deletions = ' '.join(delete_with_rm)
     cmd = f'rm -f -- {deletions}'
     log.debug('Executing: %s', cmd)
@@ -188,7 +164,9 @@ def main() -> int:
     """Entry point."""
     parser = argparse.ArgumentParser()
     parser.add_argument('-o', '--output-directory', default='.')
-    return asyncio.run(_main(parser.parse_args().output_directory))
+    parser.add_argument('-d', '--debug', action='store_true')
+    args = parser.parse_args()
+    return asyncio.run(_main(args.output_directory, debug=args.debug))
 
 
 if __name__ == "__main__":
